requirements.txt
progressbar2==3.38.0
python-utils==2.3.0
six==1.11.0
urllib3==1.23
